a) Re-use Header/RecordHeader from the client module to avoid duplication; b) Fix typos; c) Move LineMessageReader in its own top-level class.
