+++ checkstyle/suppressions.xml
@@ -267,7 +267,7 @@
     <suppress checks="BooleanExpressionComplexity"
               files="StreamsResetter.java"/>
     <suppress checks="NPathComplexity"
-              files="(ProducerPerformance|StreamsResetter|Agent|TransactionalMessageCopier|ConsoleProducer).java"/>
+              files="(ProducerPerformance|StreamsResetter|Agent|TransactionalMessageCopier|LineMessageReader).java"/>
     <suppress checks="ImportControl"
               files="SignalLogger.java"/>
     <suppress checks="IllegalImport"
